+++ src/main/java/com/example/demo/src/review/model/GetReviews.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 public class GetReviews {
     // 거래 후기 개수, List<작성자 id, 별점, 작성자 이름, 리뷰 내용, 리뷰 이미지, 리뷰 작성일, 거래 상품 id, 상품 이름>
     private int userIdx;
-    private int star;
+    private float star;
     private String userName;
     private String content;
     private String reviewImgUrl;
